% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_functions.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic knapsack}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{The data containing \code{w}, \code{v} representing the objects}

\item{W}{The weight limit of the knapsack}
}
\value{
The indices of the objects to select
}
\description{
Dynamic knapsack
}
\examples{
data(knapsack_objects)
dynamic_knapsack(knapsack_objects[1:12, ], 3500)

}
