% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_functions.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{The data containing \code{w}, \code{v} representing the objects}

\item{W}{The weight limit of the knapsack}
}
\value{
a \code{list} containing the \code{elements} and the best \code{value}
}
\description{
Greedy Knapsack
}
\examples{
data(knapsack_objects)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)

}
