% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_functions.R
\name{greedy_knapsack2}
\alias{greedy_knapsack2}
\title{Greedy Knapsack Improved}
\usage{
greedy_knapsack2(x, W)
}
\arguments{
\item{x}{A \code{data.frame} containing \code{w}, \code{v} representing the weights and values for each item}

\item{W}{The weight limit of the knapsack}
}
\value{
a \code{list} containing the \code{elements} and the best \code{value}
}
\description{
Greedy Knapsack Improved
}
\examples{
data(knapsack_objects)
greedy_knapsack2(x = knapsack_objects[1:800,], W = 3700)

}
