% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_functions.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{The data containing \code{w}, \code{v} representing the objects}

\item{W}{The weight limit of the knapsack}
}
\value{
A list containing the \code{elements} and the best \code{value}
}
\description{
Dynamic knapsack
}
\examples{
data(knapsack_objects)
knapsack_dynamic(knapsack_objects[1:12, ], 3500)

}
