% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knapsack_function}
\alias{knapsack_function}
\title{Generating data for knapsack problem with optional number of objects}
\usage{
knapsack_function(n = 2000)
}
\arguments{
\item{n}{the number of objects}
}
\value{
\code{data.frame} containing the \code{w} and \code{v}
}
\description{
Generating data for knapsack problem with optional number of objects
}
\examples{

knapsack_function(n = 666)
}
