% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{A random generated data for knapsack problem}
\format{
A data frame with 2000 rows and 2 columns:
\describe{
\item{w}{The weight of the object ranging from 1 to 4000, Integer}
\item{v}{The value of the object ranging from 0 to 10000, Numeric}
}
}
\usage{
knapsack_objects
}
\description{
This data has 2 columns, the \code{w} and \code{v}. Each row represents the weight and value of an object.
}
\keyword{datasets}
